PLATE-78 simplify code
